requirements.txt
streamlit==1.41.1
opencv-python-headless==4.10.0.84
anthropic==0.42.0
twilio==9.4.3
psycopg2-binary==2.9.10
numpy==2.2.2
python-dotenv==1.0.1
Pillow==11.1.0
